tracker/requirements.txt
Django==4.2.4
django-bootstrap5==23.3
django-filter==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==20.1.0